refactor: 불필요한 클래스 없애기
