+++ src/main/java/lotto/controller/view/OutputView.java
@@ -1,4 +1,4 @@
-package lotto.ui.view;
+package lotto.controller.view;
 
 import lotto.domain.*;
